+++ src/utils/decorators/input/__spec__.js
@@ -2,6 +2,8 @@ import React from 'react';
 import TestUtils from 'react/lib/ReactTestUtils';
 import Input from './input';
 import Icon from './../../../components/icon';
+import Help from './../../../components/help';
+import { shallow } from 'enzyme';
 
 class TestClassOne extends React.Component {
   get mainClasses() {
@@ -31,9 +33,10 @@ class TestClassTwo extends React.Component {
 
 let ExtendedClassOne = Input(TestClassOne);
 let ExtendedClassTwo = Input(TestClassTwo);
+let klass = new ExtendedClassOne;
 
 describe('Input', () => {
-  let instance, instanceTwo, onChange;
+  let instance, instanceTwo, onChange, inputHelp;
 
   beforeEach(() => {
     instance = TestUtils.renderIntoDocument(React.createElement(ExtendedClassOne, {
@@ -306,6 +309,24 @@ describe('Input', () => {
       }));
       expect(instance.iconHTML.props.className).toEqual('common-input__input-icon');
       expect(instance.iconHTML.props.children).toEqual(<Icon type='foo' />);
+    });
+  });
+
+  describe('inputHelpHTML', () => {
+    beforeEach(() => {
+      klass.props = {
+        inputHelp: 'Here is some help'
+      };
+      inputHelp = shallow(klass.inputHelpHTML);
+    });
+
+    it('returns a div with a icon', () => {
+      expect(inputHelp.props().className).toEqual('carbon-help common-input__input-help');
+      let help = inputHelp.props().children;
+      expect(help.props.tooltipMessage).toEqual('Here is some help');
+      expect(help.props.tooltipPosition).toEqual('top');
+      expect(help.props.tooltipAlign).toEqual('center');
+      expect(help.props.type).toEqual('help');
     });
   });